+++ CHANGELOG.md
@@ -28,3 +28,9 @@
 ## [v2.0.1](https://github.com/wz2cool/mybatis-dynamic-query/tree/v2.0.1) (2017-09-11)
 **Feature**
 - support serialize DynamicQuery to json
+
+## [v2.0.2](https://github.com/wz2cool/mybatis-dynamic-query/tree/v2.0.2) (2017-10-05)
+**Feature**
+- add custom sort descriptor
+- get query column ([tableName].[columnName])
+- createInstance for MybatisQueryProvider